fix(temporal): move buildPrBody out of the workflow webpack bundle path
#685's refactor put buildPrBody in `activities/docs-groom-pr.ts` and
imported it from the workflow. Temporal's workflow bundler walks every
import transitively and tries to webpack the result; docs-groom-pr.ts
imports captureWithContext / jsonLog from docs-groom-impl.ts, which
imports `@sentry/bun`, which transitively pulls Sentry's node-core
runtime that uses `node:util`, `node:worker_threads`, `node:zlib`
schemes. Webpack can't resolve those and the worker crashed at
startup:

    ERROR in node:util
    Module build failed: UnhandledSchemeError: Reading from "node:util"
      is not handled by plugins (Unhandled scheme).
    @ ./node_modules/@sentry/node-core/build/esm/integrations/systemError.js
    @ ./src/activities/docs-groom-impl.ts
    @ ./src/activities/docs-groom-pr.ts
    @ ./src/workflows/docs-groom.ts

Worker pod was in CrashLoopBackoff (7 restarts on the live cluster).

Move `buildPrBody` (and its `BuildPrBodyInput` type) to
`packages/temporal/src/shared/docs-groom-pr-body.ts` — a pure module
with no Sentry/observability imports, just the `GroomTask` type from
`docs-groom-types.ts`. Workflow imports it via `#shared/docs-groom-pr-body.ts`.
The webpack bundle no longer pulls in `@sentry/bun` and the worker
boots cleanly.

Tests still pass — the test file just updates its import path. No
behavior change vs #685's `buildPrBody` body, only its location.
